fix(theme-builder): dev error
