+++ .npmrc
@@ -1,2 +1,4 @@
 ## force pnpm to hoist
 shamefully-hoist=true
+strict-peer-dependencies=false
+node-linker=hoisted
