+++ packages/react/package.json
@@ -20,8 +20,6 @@
     "access": "public"
   },
   "scripts": {
-    "dev": "vite --host",
-    "serve": "vite preview",
     "lint": "eslint src --ext .mjs,.js,.jsx,.ts,.tsx",
     "lint:fix": "npm run lint -- --fix",
     "build": "tsc --noEmit && ./extractMuiVersion.sh && vite build",
@@ -79,7 +77,7 @@
     "react": "~19.1.0",
     "react-colorful": "^5.6.1",
     "react-dom": "~19.1.0",
-    "react-router-dom": "^7.6.3",
+    "react-router-dom": "^6.22.3",
     "recharts": "^2.15.4",
     "ufo": "^1.6.1",
     "webfontloader": "^1.6.28",
