+++ packages/react/src/constants/index.ts
@@ -1,10 +1,19 @@
+import { BLOCKLET_THEME_DARK, BLOCKLET_THEME_LIGHT } from '@blocklet/theme';
 import { type ThemeOptions } from '@mui/material';
+import { Theme, createTheme } from '@mui/material';
 
 export const PREVIEW_TEXT = 'The five boxing wizards jump quickly together';
 
 // PC 最小预览窗口宽度
 export const PC_PREVIEW_WINDOW_MIN_WIDTH = 960;
 
+// 默认浅色主题
+export const defaultLightTheme: Theme = createTheme(BLOCKLET_THEME_LIGHT);
+
+// 默认深色主题
+export const defaultDarkTheme = createTheme(BLOCKLET_THEME_DARK);
+
+// Editor 样式
 export const BUILDER_THEME_OPTIONS: ThemeOptions = {
   shape: {
     borderRadius: 8, // 固定圆角大小
