+++ packages/theme-builder/package.json
@@ -14,17 +14,15 @@
   "files": [
     "dist"
   ],
+  "main": "dist/index.js",
+  "module": "dist/index.js",
   "scripts": {
     "dev": "vite --host",
     "serve": "vite preview",
     "lint": "eslint src --ext .mjs,.js,.jsx,.ts,.tsx",
     "lint:fix": "npm run lint -- --fix",
-    "clean": "node scripts/build-clean.mjs",
-    "build": "./extractMuiVersion.sh && vite build",
+    "build": "vite build",
     "prepare": "npx simple-git-hooks",
-    "bump-version": "zx --quiet scripts/bump-version.mjs",
-    "fetch-fonts": "node scripts/fetch-google-fonts.mjs",
-    "update:deps": "taze -r -w && pnpm install && pnpm dedupe && pnpm build",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:ui": "vitest --ui",
