+++ packages/theme-builder/src/App.tsx
@@ -1,15 +1,8 @@
-import { LocaleProvider } from '@arcblock/ux/lib/Locale/context';
-
 import './app.css';
-import { translations } from './locales';
 import HomePage from './pages/index';
 
 function App() {
-  return (
-    <LocaleProvider locale="en" translations={translations}>
-      <HomePage />
-    </LocaleProvider>
-  );
+  return <HomePage />;
 }
 
 export default App;
