+++ packages/theme-builder/src/app.css
@@ -1,0 +1,12 @@
+/* Set body height to view port height, let child elements handle scroll */
+html,
+body {
+  height: 100%;
+  width: 100%;
+  overflow: hidden;
+}
+
+#app {
+  height: 100%;
+  width: 100%;
+}
