--- packages/theme-builder/src/locales/en.ts
@@ -1,183 +1,0 @@
-export default {
-  common: {
-    notifications: 'Notifications',
-    nft: 'NFTs',
-    setting: 'Settings',
-    storageManagement: 'DID Spaces',
-  },
-  editor: {
-    save: 'Save',
-    cancel: 'Cancel',
-    confirm: 'Confirm',
-    reset: 'Reset',
-    copy: 'Copy',
-    copied: 'Copied!',
-    shuffle: 'Shuffle',
-    lock: 'Lock',
-    unlock: 'Unlock',
-    undo: 'Undo',
-    redo: 'Redo',
-    concept: {
-      name: 'Theme Name',
-      add: 'Add Theme',
-      rename: 'Rename Theme',
-      duplicateName: 'Theme name already exists',
-      duplicate: 'Duplicate Theme',
-      delete: 'Delete Theme',
-      shuffle: 'Shuffle Theme',
-      saveSuccess: 'Theme saved successfully. Please refresh the page to apply changes.',
-      saveFailed: 'Save failed: {message}',
-      resetTitle: 'Reset Theme',
-      resetConfirm: 'Are you sure you want to reset the theme? This will discard all unsaved changes.',
-      resetSuccess: 'Theme has been reset to default settings. Please refresh the page to apply changes.',
-      resetFailed: 'Reset failed: {message}',
-    },
-    colorSection: {
-      title: 'Colors',
-      shades: 'Shades',
-      reset: 'Reset Colors',
-      invalidHexColor: 'Invalid hex color',
-      neutral: {
-        title: 'Neutrals',
-        default: 'background',
-        paper: 'surface',
-        divider: 'border',
-        hint: 'placeholder',
-        disabled: 'disabled text',
-        secondary: 'subtext',
-        primary: 'main text',
-      },
-      primary: 'Primary',
-      secondary: 'Secondary',
-      success: 'Success',
-      error: 'Error',
-      info: 'Info',
-      warning: 'Warning',
-      light: 'Light',
-      dark: 'Dark',
-      contrastText: 'Contrast Text',
-      lightMode: 'Light Mode',
-      darkMode: 'Dark Mode',
-      modeDisable: {
-        none: 'Allow Both',
-        dark: 'Disable Dark',
-        light: 'Disable Light',
-      },
-    },
-    typographySection: {
-      title: 'Typography',
-      reset: 'Reset All Fonts',
-      heading: 'Heading',
-      body: 'Body',
-      base: 'Base',
-      activeFont: 'Active font',
-      results: '{count} results',
-      noFontsFound: 'No matching fonts found',
-      searchFonts: 'Search fonts...',
-      sansSerif: 'Sans',
-      serif: 'Serif',
-      display: 'Display',
-      monospace: 'Mono',
-      fontSize: 'Font Size',
-      adjustFontSize: 'Adjust Font Size',
-      resetAllSizes: 'Reset All Sizes',
-      fontSizeDescriptions: {
-        body: 'Main text for content and interface',
-        h1: 'Largest titles, like page or hero headers',
-        h2: 'Big section titles',
-        h3: 'Subsection titles',
-        h4: 'Smaller titles for components',
-        h5: 'Small headings or labels',
-        h6: 'Smallest heading, for compact areas',
-        subtitle1: 'Larger secondary text',
-        subtitle2: 'Smaller secondary text',
-        overline: 'Uppercase labels above content',
-      },
-    },
-    stylesSection: {
-      title: 'Styles',
-      borderRadius: 'Border Radius',
-    },
-  },
-  samples: {
-    title: 'Samples',
-    dashboard: {
-      nav1: 'Dashboard',
-      nav2: 'DID Connect',
-      group1: 'Website',
-      group2: 'Admin',
-      group1_1: 'Overview',
-      group1_2: 'Members',
-      group1_3: 'Passports',
-      group1_4: 'Domains',
-      group2_1: 'Operations',
-      group2_2: 'Integrations',
-      group2_3: 'Developers',
-      group2_4: 'Settings',
-      content: {
-        tabs: {
-          overview: 'Overview',
-          blocklets: 'Blocklets',
-        },
-        overview: {
-          blocklets: 'Blocklets',
-          running: 'running',
-          members: 'Members',
-          active: 'Active',
-          passports: 'Passports',
-          domains: 'Domains',
-          customDomain: 'Custom domain',
-          uptime: 'Uptime',
-          createdAt: 'Created At',
-          lastBackup: 'Last Backup Time',
-        },
-        basicInfo: {
-          title: 'Basic Information',
-          appDid: 'App DID',
-          permanentDid: 'Permanent DID',
-          owner: 'Owner',
-          installedAt: 'Installed At',
-          walletType: 'Wallet Type',
-          serverDid: 'Server DID',
-          serverVersion: 'Server Version',
-        },
-      },
-    },
-    website: {
-      hero: {
-        title: 'Build Better Teams with',
-        subtitle:
-          'The ultimate collaboration platform for modern IT teams. Streamline your workflow, enhance communication, and deliver projects faster.',
-        cta: {
-          start: 'Get Started Free',
-          demo: 'Watch Demo',
-        },
-      },
-      features: {
-        title: 'Why Teams Choose TeamSpace',
-        subtitle: 'Powerful features designed for modern development teams',
-        collaboration: {
-          title: 'Team Collaboration',
-          description: 'Real-time collaboration tools that keep your team in sync',
-        },
-        performance: {
-          title: 'Fast Performance',
-          description: 'Lightning-fast performance that scales with your team',
-        },
-        security: {
-          title: 'Enterprise Security',
-          description: 'Bank-grade security to protect your sensitive data',
-        },
-        access: {
-          title: 'Global Access',
-          description: 'Access your workspace from anywhere in the world',
-        },
-      },
-      finalCta: {
-        title: 'Ready to Transform Your Team?',
-        subtitle: 'Join thousands of teams already using TeamSpace to build better products.',
-        cta: 'Start Your Free Trial',
-      },
-    },
-  },
-};
