--- packages/theme-builder/src/locales/index.ts
@@ -1,9 +1,0 @@
-import { flatten } from 'flat';
-
-import en from './en';
-import zh from './zh';
-
-export const translations = {
-  en: flatten(en),
-  zh: flatten(zh),
-};
